Revert "feat(gcl/cdl): add comprehensive guide page with enterprise best practices"
This reverts commit 92e844a4fd88914f74af5ad8bc938e4b495f2d4d.
